Run hourly, fix no img when tweetsArray has 1
Now instead of running a different scheduler in Gcloud for each country,
run the scheduler hourly and see if it's 13, 18 or 23 o'clock in any of
the countries. To avoid hour to be posted as 13:03 unrounded, get the
date immediately at the function start and pass to tweeter from there.

Fix country changing when iterating over config keys with const country.

Fix summary image not being posted when tweetArray.length === 1. Remove
the line.